+++ tests/integration/networking_libraries/test_requests_client.py
@@ -11,14 +11,8 @@ import pyodata.v2.service
 import pyodata.v3.service
 from pyodata.exceptions import PyODataException, HttpError
 from pyodata.v2.model import ParserError, PolicyWarning, PolicyFatal, PolicyIgnore, Config
-from tests.conftest import contents_of_fixtures_file
 
 SERVICE_URL = 'http://example.com'
-
-
-@pytest.fixture
-def metadata_v3():
-    return contents_of_fixtures_file('metadata_v3.xml')